Fix data type of semi targets

+++ dataset.py
@@ -71,10 +71,10 @@ class LGDataset(Dataset):
 
         if self.train:
             self.data = X_train
-            self.targets = torch.tensor(y_train)
+            self.targets = torch.tensor(y_train, dtype=torch.int32)
         else:
             self.data = X_test
-            self.targets = torch.tensor(y_test)
+            self.targets = torch.tensor(y_test, dtype=torch.int32)
 
         self.semi_targets = torch.zeros_like(self.targets)
